moved tests and test utils to test dir
