--- src/utils/test/index.ts
@@ -1,2 +1,0 @@
-export * from './callGraphQL'
-export * from './testConnection'
